// src/types.ts
export type Maybe<T> = T | null;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  ObjectId: any;
};
// ...
export type Project = {
  __typename?: 'Project';
  _id?: Maybe<Scalars['ObjectId']>;
  name?: Maybe<Scalars['String']>;
  users?: Maybe<Array<Maybe<User>>>;
};
// ...
export type Task = {
  __typename?: 'Task';
  _id?: Maybe<Scalars['ObjectId']>;
  assignee?: Maybe<User>;
  description: Scalars['String'];
  status: TaskStatus;
};
// ...
export enum TaskStatus {
  Complete = 'COMPLETE',
  Open = 'OPEN',
  Inprogress = 'INPROGRESS'
}
// ...
export type User = {
  __typename?: 'User';
  _id?: Maybe<Scalars['ObjectId']>;
  image?: Maybe<Scalars['String']>;
  name: Scalars['String'];
  projects?: Maybe<Array<Maybe<Project>>>;
  user_id: Scalars['String'];
};